config fixes

+++ deployment/indexers-mainnet/stop.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
 chain="$1"
+spammer="$2"
 
-docker-compose -f docker-compose-indexer-$chain.yaml -p indexer-$chain down
+if [[ $spammer == spammer ]] ; then
+   docker-compose -f docker-compose-indexer-$chain.yaml -p indexer-$chain stop spammer 
+else
+   docker-compose -f docker-compose-indexer-$chain.yaml -p indexer-$chain down
+fi
